clear code, move policy functions to prepare.tx, add utils.ts for tool functions.
